breakup qt/_layers_list
